completed user restriction

+++ base/views.py
@@ -1,13 +1,17 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(req):
     context = {}
+    if req.user.is_authenticated:
+        return redirect('home')
     if req.method == "POST":
         username = req.POST.get('username')
         password = req.POST.get('password')
@@ -44,6 +48,7 @@ def room(req, pk):
     context = {'room':room}
     return render(req, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(req):
     form = RoomForm()
 
@@ -55,9 +60,13 @@ def createRoom(req):
     context = {'form' : form}
     return render(req, 'base/room_form.html', context )
 
+@login_required(login_url='login')
 def updateRoom(req, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance=room)
+
+    if req.user != room.host:
+        return HttpResponse('You are not allowed here. ')
 
     if req.method == 'POST':
         form = RoomForm(req.POST, instance=room)
@@ -67,8 +76,11 @@ def updateRoom(req, pk):
     context = {'form' : form }
     return render(req, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(req, pk):
     room = Room.objects.get(id=pk)
+    if req.user != room.host:
+        return HttpResponse('You are not allowed here. ')
     if req.method == 'POST':
         room.delete()
         return redirect('home')
